Add TvOS support to the auth module.
There are some parts of the public api, that are not existing in the TvOS sdk.
Mostly relating to auth via phone number or multifactor.
Those methods will throw an UnsupportedOperationException with this change.
