+++ firebase-auth/src/appleMain/kotlin/dev/gitlive/firebase/auth/auth.kt
@@ -8,8 +8,6 @@ import cocoapods.FirebaseAuth.*
 import dev.gitlive.firebase.Firebase
 import dev.gitlive.firebase.FirebaseApp
 import dev.gitlive.firebase.FirebaseException
-import dev.gitlive.firebase.FirebaseNetworkException
-import dev.gitlive.firebase.FirebaseTooManyRequestsException
 import dev.gitlive.firebase.auth.ActionCodeResult.*
 import dev.gitlive.firebase.ios
 import kotlinx.cinterop.*
@@ -166,10 +164,11 @@ public actual open class FirebaseAuthEmailException(message: String) : FirebaseA
 public actual open class FirebaseAuthInvalidCredentialsException(message: String) : FirebaseAuthException(message)
 public actual open class FirebaseAuthWeakPasswordException(message: String) : FirebaseAuthInvalidCredentialsException(message)
 public actual open class FirebaseAuthInvalidUserException(message: String) : FirebaseAuthException(message)
-public actual open class FirebaseAuthMultiFactorException(message: String, public val resolver: FIRMultiFactorResolver?) : FirebaseAuthException(message)
 public actual open class FirebaseAuthRecentLoginRequiredException(message: String) : FirebaseAuthException(message)
 public actual open class FirebaseAuthUserCollisionException(message: String) : FirebaseAuthException(message)
 public actual open class FirebaseAuthWebException(message: String) : FirebaseAuthException(message)
+
+internal expect fun NSError.toAuthException(): FirebaseException
 
 internal fun <T, R> T.throwError(block: T.(errorPointer: CPointer<ObjCObjectVar<NSError?>>) -> R): R {
     memScoped {
@@ -177,7 +176,7 @@ internal fun <T, R> T.throwError(block: T.(errorPointer: CPointer<ObjCObjectVar<
         val result = block(errorPointer)
         val error: NSError? = errorPointer.pointed.value
         if (error != null) {
-            throw error.toException()
+            throw error.toAuthException()
         }
         return result
     }
@@ -189,7 +188,7 @@ internal suspend inline fun <T, reified R> T.awaitResult(function: T.(callback: 
         if (error == null) {
             job.complete(result)
         } else {
-            job.completeExceptionally(error.toException())
+            job.completeExceptionally(error.toAuthException())
         }
     }
     return job.await() as R
@@ -201,70 +200,8 @@ internal suspend inline fun <T> T.await(function: T.(callback: (NSError?) -> Uni
         if (error == null) {
             job.complete(Unit)
         } else {
-            job.completeExceptionally(error.toException())
+            job.completeExceptionally(error.toAuthException())
         }
     }
     job.await()
-}
-
-private fun NSError.toException() = when (domain) {
-    // codes from AuthErrors.swift: https://github.com/firebase/firebase-ios-sdk/blob/
-    // 2f6ac4c2c61cd57c7ea727009e187b7e1163d613/FirebaseAuth/Sources/Swift/Utilities/
-    // AuthErrors.swift#L51
-    FIRAuthErrorDomain -> when (code) {
-        17030L, // AuthErrorCode.invalidActionCode
-        17029L, // AuthErrorCode.expiredActionCode
-        -> FirebaseAuthActionCodeException(toString())
-
-        17008L, // AuthErrorCode.invalidEmail
-        -> FirebaseAuthEmailException(toString())
-
-        17056L, // AuthErrorCode.captchaCheckFailed
-        17042L, // AuthErrorCode.invalidPhoneNumber
-        17041L, // AuthErrorCode.missingPhoneNumber
-        17046L, // AuthErrorCode.invalidVerificationID
-        17044L, // AuthErrorCode.invalidVerificationCode
-        17045L, // AuthErrorCode.missingVerificationID
-        17043L, // AuthErrorCode.missingVerificationCode
-        17021L, // AuthErrorCode.userTokenExpired
-        17004L, // AuthErrorCode.invalidCredential
-        -> FirebaseAuthInvalidCredentialsException(toString())
-
-        17026L, // AuthErrorCode.weakPassword
-        -> FirebaseAuthWeakPasswordException(toString())
-
-        17017L, // AuthErrorCode.invalidUserToken
-        -> FirebaseAuthInvalidUserException(toString())
-
-        17014L, // AuthErrorCode.requiresRecentLogin
-        -> FirebaseAuthRecentLoginRequiredException(toString())
-
-        17087L, // AuthErrorCode.secondFactorAlreadyEnrolled
-        17078L, // AuthErrorCode.secondFactorRequired
-        17088L, // AuthErrorCode.maximumSecondFactorCountExceeded
-        17084L, // AuthErrorCode.multiFactorInfoNotFound
-        -> {
-            val resolver = userInfo["FIRAuthErrorUserInfoMultiFactorResolverKey"] as? FIRMultiFactorResolver
-            FirebaseAuthMultiFactorException(toString(), resolver)
-        }
-
-        17052L, // AuthErrorCode.quotaExceeded
-        -> FirebaseTooManyRequestsException(toString())
-
-        17007L, // AuthErrorCode.emailAlreadyInUse
-        17012L, // AuthErrorCode.accountExistsWithDifferentCredential
-        17025L, // AuthErrorCode.credentialAlreadyInUse
-        -> FirebaseAuthUserCollisionException(toString())
-
-        17057L, // AuthErrorCode.webContextAlreadyPresented
-        17058L, // AuthErrorCode.webContextCancelled
-        17062L, // AuthErrorCode.webInternalError
-        -> FirebaseAuthWebException(toString())
-
-        17020L, // AuthErrorCode.networkError
-        -> FirebaseNetworkException(toString())
-
-        else -> FirebaseAuthException(toString())
-    }
-    else -> FirebaseAuthException(toString())
 }
