+++ firebase-auth/src/tvosMain/kotlin/dev/gitlive/firebase/auth/user.kt
@@ -1,0 +1,5 @@
+package dev.gitlive.firebase.auth
+
+internal actual fun FirebaseUser.getMultiFactor(): MultiFactor = MultiFactor()
+
+internal actual suspend fun FirebaseUser.updatePhoneNumberInternal(credential: PhoneAuthCredential): Unit = Unit
