+++ gradle.properties
@@ -21,7 +21,7 @@ kotlin.native.cacheKind=none
 # supportedTargets=ios,macos,tvos,jvm,js,android
 firebase-analytics.supportedTargets=ios,macos,tvos,jvm,js,android
 firebase-app.supportedTargets=ios,macos,tvos,jvm,js,android
-firebase-auth.supportedTargets=ios,jvm,js,android
+firebase-auth.supportedTargets=ios,tvos,jvm,js,android
 firebase-common-internal.supportedTargets=ios,macos,tvos,jvm,js,android
 firebase-common.supportedTargets=ios,macos,tvos,jvm,js,android
 firebase-config.supportedTargets=ios,macos,tvos,jvm,js,android
