+++ gradle/libs.versions.toml
@@ -8,7 +8,7 @@ firebase-bom = "33.15.0"
 gitlive-firebase-java-sdk = "0.6.3"
 gson = "2.13.1"
 junit = "4.13.2"
-kotlin = "2.2.20"
+kotlin = "2.2.21"
 kotlinx-coroutines = "1.10.2"
 kotlinx-serialization = "1.9.0"
 kotlinx-binarycompatibilityvalidator = "0.18.1"
